THE ASSOCIATION OF TRAINERS, CAREER COUNSELORS & COACHES!
Our aim is to help youth, to lift them up from college level to corporate level and from unemployment to entrepreneurship.
We work collaboratively with different Educational & Skill Development Organizations and conduct different sessions for Learners and for Teaching & Non-Teaching staff.
We are passionate about improving the lives of our own with the people of the world through our training, speaking, counseling and coaching within the country and around the globe. We inspire and encourage people by making them realize the true potential.